# contracts/web-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "Cherry OJ Web API",
    "version": "0.2.0",
    "description": "浏览器与 Gateway 之间的公开 REST 契约。请求使用 endpoint DTO，普通 JSON 成功响应使用 ApiSuccess，失败响应使用 RFC 9457 Problem Details。 管理部署接口的 503 可返回固定节点错误：NO_ONLINE_JUDGE_NODE, JUDGE_NODE_UNREACHABLE, JUDGE_NODE_DATA_REJECTED, JUDGE_NODE_RECEIPT_MISMATCH。详情由固定文案生成，不透传节点正文；readiness checks 的 ONLINE_JUDGE_NODE 表示 ACTIVE 环境存在在线节点，DEPLOYMENT 仅在该在线节点持有匹配摘要回执时通过。"
  },
...
  "paths": {
...
    "/api/problems/{slug}": {
      "get": {
        "operationId": "getProblem",
        "summary": "读取公开题目当前版本",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemSlug"
          }
        ],
        "responses": {
          "200": {
            "description": "公开题目详情；不包含隐藏数据、判题模板、作者或审计字段",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailSuccess"
                },
                "example": {
                  "data": {
                    "problemId": "019c8e42-7f70-7000-8000-000000000101",
                    "slug": "a-plus-b",
                    "title": "A+B Problem",
                    "difficulty": "EASY",
                    "tags": [
                      "入门"
                    ],
                    "codeMode": "ACM",
                    "statementMarkdown": "计算两个整数之和。",
                    "inputDescriptionMarkdown": "输入两个整数。",
                    "outputDescriptionMarkdown": "输出它们的和。",
                    "constraintsMarkdown": "整数在 32 位有符号范围内。",
                    "hintMarkdown": null,
                    "samples": [
                      {
                        "ordinal": 1,
                        "input": "1 2\n",
                        "output": "3\n",
                        "explanationMarkdown": null
                      }
                    ],
                    "allowedLanguages": [
                      {
                        "id": "cpp",
                        "displayName": "C++",
                        "starterCode": "#include <iostream>\n"
                      }
                    ]
                  },
                  "meta": {
                    "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
...
    "/api/admin/problems": {
      "get": {
        "operationId": "listAdminProblems",
        "summary": "分页读取题目与版本状态",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProblemStatus"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "题目管理分页",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemListSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      },
      "post": {
        "operationId": "createProblem",
        "summary": "创建 C++ ACM 题目及初始草稿",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProblemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "题目及 v1 草稿已创建",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Location": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/admin/problems/{problemId}": {
      "get": {
        "operationId": "getAdminProblem",
        "summary": "读取题目管理详情",
        "description": "题目只有一份内容：题面、样例、语言模板与测试数据摘要都在这里。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          }
        ],
        "responses": {
          "200": {
            "description": "题目管理模型",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      },
      "patch": {
        "operationId": "updateProblem",
        "summary": "保存题目",
        "description": "改了就是改了，没有草稿或版本。公开题目的修改立即对学生生效，因此不允许把公开题目改成空题面或没有样例；旧 rowVersion 返回 409 ROW_VERSION_CONFLICT。可见性与测试数据由专门的接口改变。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProblemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "题目已保存",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      },
      "delete": {
        "operationId": "deleteProblem",
        "summary": "删除从未公开过的题目",
        "description": "只允许删除从未公开过的题目（它不可能有提交）；题面、样例、语言、审计事件和测试数据目录一并删除。公开过的题目只能归档，返回 409 RESOURCE_STATE_CONFLICT。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/RowVersion"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContentResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/admin/problems/{problemId}/archive": {
      "post": {
        "operationId": "archiveProblem",
        "summary": "归档题目并从公开题库移除",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "题目已归档",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/admin/problems/{problemId}/test-data": {
      "get": {
        "operationId": "getTestData",
        "summary": "读取题目当前的测试数据",
        "description": "指纹、测试点数和文件清单来自测试数据目录里的 testdata.json（协议见 docs/testdata-protocol.md）；不暴露服务器上的目录地址。没有测试数据返回 404 TEST_DATA_NOT_FOUND。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          }
        ],
        "responses": {
          "200": {
            "description": "当前测试数据",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestDataSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      },
      "put": {
        "operationId": "replaceTestData",
        "summary": "上传 ZIP，替换题目的测试数据",
        "description": "题目只有一份测试数据，上传即整体替换。ZIP 的逻辑根只允许成对的 <case>.in/.out UTF-8 普通文件；文件可直接位于根目录或统一位于一个外层文件夹。已知 macOS 元数据（.DS_Store、__MACOSX、._*）会被忽略。校验通过后写成协议目录并原子切换，非法 ZIP 不改变已有数据。数据换了之后，旧的校准过期，需要重新校准才能继续接收新提交。网关和下游按配置限额流式处理；超限返回 413 PAYLOAD_TOO_LARGE。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadTestDataRequest"
              },
              "encoding": {
                "file": {
                  "contentType": "application/zip"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "测试数据已替换",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestDataSuccess"
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/ApiProblemResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
...
    "/api/admin/problems/{problemId}/preview": {
      "get": {
        "operationId": "previewProblem",
        "summary": "按公开题面渲染预览",
        "description": "返回学生会看到的题面形态；私有题目也可预览。响应不得缓存。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          }
        ],
        "responses": {
          "200": {
            "description": "题面预览",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/admin/problems/{problemId}/calibration": {
      "post": {
        "operationId": "calibrateProblem",
        "summary": "以临时 C++ 参考程序验证绝对限制",
        "description": "参考源码仅用于本次同步验证，不进入响应、业务表、日志或审计正文。只有当前测试数据全 AC 才产生 VALID 校准；校准记录它所用测试数据的指纹，数据更新后即过期。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrateProblemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "校准尝试完成；verdict 非 AC 时 status 为 FAILED 且不会覆盖已有 VALID 事实",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageCalibrationSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/admin/problems/{problemId}/publish-check": {
      "get": {
        "operationId": "checkProblemPublication",
        "summary": "检查题目能否公开",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          }
        ],
        "responses": {
          "200": {
            "description": "六项检查：内容、样例、语言、测试数据、校准、在线判题节点",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishCheckSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/admin/problems/{problemId}/publish": {
      "post": {
        "operationId": "publishProblem",
        "summary": "公开题目",
        "description": "检查全部通过才公开；检查之后测试数据被替换则拒绝。重试读取并返回相同发布事实，并发冲突返回 409。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishProblemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "题目已公开",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/admin/problems/{problemId}/unpublish": {
      "post": {
        "operationId": "unpublishProblem",
        "summary": "取消公开",
        "description": "题目变回私有；已有提交记录保留。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "题目已取消公开",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "CHERRY_SESSION",
        "description": "浏览器随机 Session Cookie；不是内部 JWT。"
      }
    },
    "parameters": {
...
      "ProblemSlug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/ProblemSlug"
        }
      },
...
    },
    "headers": {
      "XRequestId": {
        "description": "Gateway 生成的公开请求标识；必须与响应 body 的 meta.requestId 相同。",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/RequestId"
        }
      }
    },
    "responses": {
...
      "ApiProblemResponse": {
        "description": "符合 RFC 9457 且经过 Gateway 脱敏的错误响应。实际 HTTP status 与 body.status 相同。",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "urn:cherry-oj:problem:validation-failed",
              "title": "请求参数校验失败",
              "status": 422,
              "detail": "请检查标记字段。",
              "instance": "urn:cherry-oj:request:req_01K37XZ3MFXBK92WMG67G4XFN0",
              "code": "VALIDATION_FAILED",
              "meta": {
                "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
              },
              "violations": [
                {
                  "path": "title",
                  "code": "SIZE",
                  "message": "标题长度必须在 1 到 120 之间。"
                }
              ]
            }
          }
        }
      }
    },
    "schemas": {
...
      "ProblemStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "ARCHIVED"
        ]
      },
...
      "ProblemDetailSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProblemDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "AdminProblemListSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminProblemListData"
          },
          "meta": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiMeta"
              }
            ],
            "required": [
              "requestId",
              "pagination"
            ]
          }
        }
      },
...
    }
  }
}